requirements-dev.txt
pytest
pytest-pretty
pytest-cov
gitpython